requirements.txt
streamlit
langchain-community
vaderSentiment
textblob
requests
python-dotenv
transformers
torch
